+++ efs/main.tf
@@ -1,5 +1,5 @@
 module "vpc_map" {
-  source                              = "../vpc_id_map"
+  source                              = "../vpc/id_map"
   vpc_map                             = var.fs_map
   vpc_az_key_list_default             = var.vpc_az_key_list_default
   vpc_key_default                     = var.vpc_key_default