+++ lambda_function/locals.tf
@@ -19,7 +19,7 @@ module "policy_map" {
 }
 
 module "vpc_map" {
-  source                              = "../vpc_id_map"
+  source                              = "../vpc/id_map"
   vpc_az_key_list_default             = var.vpc_az_key_list_default
   vpc_data_map                        = var.vpc_data_map
   vpc_key_default                     = var.vpc_key_default